+++ src/services/mod.rs
@@ -10,6 +10,7 @@ pub mod mpris;
 pub mod network;
 pub mod notifications;
 pub mod privacy;
+pub mod tempo_calendar;
 mod throttle;
 pub mod tray;
 pub mod upower;
